requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-telegram-bot==22.1
discord.py==2.5.2
mistralai==1.8.2
pdfplumber==0.11.7
reportlab==4.2.5
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.7
email-validator==2.2.0
python-dotenv==1.0.1
